Use translucent black for RoomSubList bg to fix tinting
The alternative is to specify yet another colour that we have to
worry about when applying tinting, which is currently quite
fragile and requires changes to many parts of the app, including
all themes that don't require the colour to be tinted.

By using translucent black, we effectively take the secondary
accent colour of the LeftPanel and make it more black.

Note: This does not preserve the previous colour we were using,
although the alternative above does allow for this.